Refactor Prompt Tests
## Changes Made

- Created a new abstract base class `AbstractPromptTest` that contains shared test fields, constants, and data providers
- Extracted serialization-specific tests into a new `PromptSerializationTest` class
- Updated the original `PromptTest` class to contain only non-serialization tests
- Both test classes now extend the abstract base class to reuse common test infrastructure
- Replaced all JUnit assertions (assertEquals, assertTrue, assertNull) with their Kotest equivalents (shouldBe, shouldBeTrue, shouldBeNull) for improved readability
- Implemented `assertSoftly(subject) { ... }` for multiple assertions on the same subject
- Added `withClue` for non-obvious assertions to provide better context for failures

## Motivation
This refactoring improves the test organization by:
- Separating concerns: Serialization tests are now isolated from other prompt functionality tests
- Reducing code duplication: Common test setup and data providers are now in a shared abstract class
- Improving maintainability: Each test class has a clear, focused responsibility
- Making tests more readable: Smaller, more focused test classes are easier to understand
- Enhancing test failure messages: Kotest assertions provide more informative error messages
- Grouping related assertions: Using `assertSoftly` to see all failures at once rather than stopping at the first failure
