+++ prompt/prompt-model/build.gradle.kts
@@ -22,6 +22,8 @@ kotlin {
         commonTest {
             dependencies {
                 implementation(kotlin("test"))
+                implementation(libs.kotest.assertions.core)
+                implementation(libs.kotest.assertions.json)
             }
         }
