+++ prompt/prompt-model/src/jvmTest/kotlin/ai/koog/prompt/PromptTest.kt
@@ -2,99 +2,17 @@ package ai.koog.prompt
 
 import ai.koog.prompt.dsl.Prompt
 import ai.koog.prompt.message.Message
-import ai.koog.prompt.message.RequestMetaInfo
-import ai.koog.prompt.message.ResponseMetaInfo
 import ai.koog.prompt.params.LLMParams
-import kotlinx.datetime.Clock
-import kotlinx.datetime.Instant
-import kotlinx.serialization.SerializationException
-import kotlinx.serialization.json.Json
+import io.kotest.assertions.assertSoftly
+import io.kotest.assertions.withClue
+import io.kotest.matchers.shouldBe
+import io.kotest.matchers.shouldNotBe
+import io.kotest.matchers.types.shouldBeInstanceOf
 import kotlinx.serialization.json.buildJsonObject
 import kotlinx.serialization.json.put
-import org.junit.jupiter.api.Assertions.assertNull
-import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.assertThrows
-import org.junit.jupiter.params.ParameterizedTest
-import org.junit.jupiter.params.provider.MethodSource
-import java.util.stream.Stream
-import kotlin.test.assertEquals
-import kotlin.test.assertNotEquals
 
-class PromptTest {
-    companion object {
-        val ts: Instant = Instant.parse("2023-01-01T00:00:00Z")
-
-        val testClock: Clock = object : Clock {
-            override fun now(): Instant = ts
-        }
-
-        val testRespMetaInfo = ResponseMetaInfo.create(testClock)
-        val testReqMetaInfo = RequestMetaInfo.create(testClock)
-
-        val promptId = "test-id"
-        val systemMessage = "You are a helpful assistant with many capabilities"
-        val assistantMessage = "I'm here to help!"
-        val userMessage = "Can you help me calculate 5 + 3?"
-        val speculationMessage = "The result is 8"
-        val toolCallId = "tool_call_123"
-        val toolName = "calculator"
-        val toolCallContent = """{"operation": "add", "a": 5, "b": 3}"""
-        val toolResultContent = "8"
-        val finishReason = "stop"
-        val emptyName = ""
-
-        val simpleSchemaName = "simple-schema"
-        val simpleSchema = buildJsonObject {
-            put("type", "string")
-        }
-
-        val fullSchemaName = "full-schema"
-        val fullSchema = buildJsonObject {
-            put("type", "object")
-            put("required", true)
-        }
-
-        val basicPrompt = Prompt.build("test", clock = testClock) {
-            system(systemMessage)
-            user(userMessage)
-            message(
-                Message.Assistant(
-                    content = assistantMessage,
-                    metaInfo = testRespMetaInfo,
-                    finishReason = finishReason
-                )
-            )
-            tool {
-                call(toolCallId, toolName, toolCallContent)
-                result(toolCallId, toolName, toolResultContent)
-            }
-        }
-
-        @JvmStatic
-        fun toolChoiceSerializationProvider(): Stream<Array<Any>> = Stream.of(
-            arrayOf("Auto", LLMParams.ToolChoice.Auto),
-            arrayOf("Required", LLMParams.ToolChoice.Required),
-            arrayOf("Named", LLMParams.ToolChoice.Named(toolName)),
-            arrayOf("None", LLMParams.ToolChoice.None)
-        )
-
-        @JvmStatic
-        fun schemaSerializationProvider(): Stream<Array<Any>> = Stream.of(
-            arrayOf(
-                "Simple JSON Schema",
-                LLMParams.Schema.JSON.Simple(simpleSchemaName, simpleSchema),
-                simpleSchemaName,
-                LLMParams.Schema.JSON.Simple::class.java
-            ),
-            arrayOf(
-                "Full JSON Schema",
-                LLMParams.Schema.JSON.Full(fullSchemaName, fullSchema),
-                fullSchemaName,
-                LLMParams.Schema.JSON.Full::class.java
-            )
-        )
-    }
+class PromptTest : AbstractPromptTest() {
 
     @Test
     fun testPromptBuilding() {
@@ -114,124 +32,27 @@ class PromptTest {
             }
         }
 
-        assertEquals(5, prompt.messages.size)
-        assertTrue(prompt.messages[0] is Message.System)
-        assertTrue(prompt.messages[1] is Message.User)
-        assertTrue(prompt.messages[2] is Message.Assistant)
-        assertTrue(prompt.messages[3] is Message.Tool.Call)
-        assertTrue(prompt.messages[4] is Message.Tool.Result)
-
-        assertEquals(systemMessage, prompt.messages[0].content)
-        assertEquals(userMessage, prompt.messages[1].content)
-        assertEquals(assistantMessage, prompt.messages[2].content)
-        assertEquals(toolContent, prompt.messages[3].content)
-        assertEquals(toolResult, prompt.messages[4].content)
-        assertEquals(toolName, (prompt.messages[3] as Message.Tool.Call).tool)
-        assertEquals(toolName, (prompt.messages[4] as Message.Tool.Result).tool)
-    }
-
-    @Test
-    fun testBasicSerialization() {
-        val json = Json.encodeToString(basicPrompt)
-        val decoded = Json.decodeFromString<Prompt>(json)
-
-        assertEquals(basicPrompt, decoded)
-        assertEquals(basicPrompt.messages.size, decoded.messages.size)
-        for (i in basicPrompt.messages.indices) {
-            assertTrue(decoded.messages[i].role == basicPrompt.messages[i].role)
-        }
-    }
-
-    @Test
-    fun testPromptSerialization() {
-        val prompt = basicPrompt.withUpdatedParams {
-            temperature = 0.7
-            speculation = speculationMessage
-            schema = LLMParams.Schema.JSON.Simple(simpleSchemaName, simpleSchema)
-            toolChoice = LLMParams.ToolChoice.Auto
-            user = "test_user"
-        }
-
-        val encodedPrompt = Json.encodeToString(prompt)
-        val decodedPrompt = Json.decodeFromString<Prompt>(encodedPrompt)
-
-        assertEquals(prompt, decodedPrompt)
-        assertEquals(prompt.messages.size, decodedPrompt.messages.size)
-        assertEquals(0.7, decodedPrompt.params.temperature)
-        assertEquals(speculationMessage, decodedPrompt.params.speculation)
-        assertTrue(decodedPrompt.params.schema is LLMParams.Schema.JSON.Simple)
-        assertEquals(simpleSchemaName, decodedPrompt.params.schema?.name)
-        assertTrue(decodedPrompt.params.toolChoice is LLMParams.ToolChoice.Auto)
-        assertEquals("test_user", decodedPrompt.params.user)
-
-        decodedPrompt.messages.forEachIndexed { index, decodedMessage ->
-            assertTrue(decodedMessage.role == prompt.messages[index].role)
-            assertTrue(decodedMessage.content == prompt.messages[index].content)
-            if (decodedMessage.role == Message.Role.Assistant) {
-                assertTrue(
-                    (decodedMessage as Message.Assistant).finishReason ==
-                            (prompt.messages[index] as Message.Assistant).finishReason
-                )
+        assertSoftly(prompt.messages) {
+            size shouldBe 5
+            get(0).shouldBeInstanceOf<Message.System> {
+                it.content shouldBe systemMessage
             }
-
-            if (decodedMessage.role == Message.Role.Tool) {
-                if (decodedMessage is Message.Tool.Call) {
-                    val originalToolMessage = prompt.messages[index] as Message.Tool.Call
-                    val decodedToolMessage = decodedMessage as Message.Tool.Call
-                    assertTrue(decodedToolMessage.id == originalToolMessage.id)
-                    assertTrue(decodedToolMessage.tool == originalToolMessage.tool)
-                    assertTrue(decodedToolMessage.content == originalToolMessage.content)
-                } else if (decodedMessage is Message.Tool.Result) {
-                    val originalToolMessage = prompt.messages[index] as Message.Tool.Result
-                    val decodedToolMessage = decodedMessage as Message.Tool.Result
-                    assertTrue(decodedToolMessage.id == originalToolMessage.id)
-                    assertTrue(decodedToolMessage.tool == originalToolMessage.tool)
-                    assertTrue(decodedToolMessage.content == originalToolMessage.content)
-                }
+            get(1).shouldBeInstanceOf<Message.User> {
+                it.content shouldBe userMessage
             }
-        }
-    }
-
-    @ParameterizedTest
-    @MethodSource("schemaSerializationProvider")
-    fun testSchemaSerialization(name: String, schema: LLMParams.Schema, schemaName: String, schemaClass: Class<*>) {
-        val prompt = basicPrompt.withUpdatedParams {
-            this.schema = schema
-        }
-
-        val schemaJson = Json.encodeToString(prompt)
-        val decodedSchema = Json.decodeFromString<Prompt>(schemaJson)
-
-        assertEquals(prompt, decodedSchema)
-        assertEquals(prompt.messages.size, decodedSchema.messages.size)
-        assertTrue(schemaClass.isInstance(decodedSchema.params.schema))
-        assertEquals(schemaName, decodedSchema.params.schema?.name)
-
-        decodedSchema.messages.forEachIndexed { index, decodedMessage ->
-            assertTrue(decodedMessage.role == prompt.messages[index].role)
-            assertTrue(decodedMessage.content == prompt.messages[index].content)
-        }
-    }
-
-    @ParameterizedTest
-    @MethodSource("toolChoiceSerializationProvider")
-    fun testToolChoiceSerialization(name: String, toolChoiceOption: LLMParams.ToolChoice) {
-        val prompt = basicPrompt.withUpdatedParams {
-            toolChoice = toolChoiceOption
-        }
-        val toolChoiceJson = Json.encodeToString(prompt)
-        val decodedToolChoice = Json.decodeFromString<Prompt>(toolChoiceJson)
-
-        assertEquals(prompt, decodedToolChoice)
-        assertEquals(prompt.messages.size, decodedToolChoice.messages.size)
-        assertTrue(decodedToolChoice.params.toolChoice == toolChoiceOption)
-        if (toolChoiceOption is LLMParams.ToolChoice.Named) {
-            assertEquals(toolName, (decodedToolChoice.params.toolChoice as LLMParams.ToolChoice.Named).name)
-        }
-
-        decodedToolChoice.messages.forEachIndexed { index, decodedMessage ->
-            assertTrue(decodedMessage.role == prompt.messages[index].role)
-            assertTrue(decodedMessage.content == prompt.messages[index].content)
+            get(2).shouldBeInstanceOf<Message.Assistant> {
+                it.content shouldBe assistantMessage
+            }
+            get(3).shouldBeInstanceOf<Message.Tool.Call> {
+                it.id shouldBe toolCallId
+                it.tool shouldBe toolName
+                it.content shouldBe toolContent
+            }
+            get(4).shouldBeInstanceOf<Message.Tool.Result> {
+                it.id shouldBe toolCallId
+                it.tool shouldBe toolName
+                it.content shouldBe toolResult
+            }
         }
     }
 
@@ -249,10 +70,12 @@ class PromptTest {
 
         val updatedPrompt = basicPrompt.withMessages { newMessages }
 
-        assertEquals(3, updatedPrompt.messages.size)
-        assertEquals(systemMessage, updatedPrompt.messages[0].content)
-        assertEquals(userMessage, updatedPrompt.messages[1].content)
-        assertEquals(assistantMessage, updatedPrompt.messages[2].content)
+        assertSoftly(updatedPrompt.messages) {
+            size shouldBe 3
+            get(0).content shouldBe systemMessage
+            get(1).content shouldBe userMessage
+            get(2).content shouldBe assistantMessage
+        }
     }
 
     @Test
@@ -272,12 +95,16 @@ class PromptTest {
 
         val updatedPrompt = basicPrompt.withParams(newParams)
 
-        assertEquals(0.7, updatedPrompt.params.temperature)
-        assertEquals(speculation, updatedPrompt.params.speculation)
-        assertTrue(updatedPrompt.params.schema is LLMParams.Schema.JSON.Simple)
-        assertEquals(schemaName, updatedPrompt.params.schema?.name)
-        assertTrue(updatedPrompt.params.toolChoice is LLMParams.ToolChoice.Auto)
-        assertEquals("test_user", updatedPrompt.params.user)
+        assertSoftly(updatedPrompt.params) {
+            temperature shouldBe 0.7
+            this.speculation shouldBe speculation
+            withClue("Schema should be of type Simple") {
+                schema.shouldBeInstanceOf<LLMParams.Schema.JSON.Simple>()
+            }
+            schema?.name shouldBe schemaName
+            toolChoice shouldBe LLMParams.ToolChoice.Auto
+            user shouldBe "test_user"
+        }
     }
 
     @Test
@@ -298,288 +125,18 @@ class PromptTest {
             user = "updated_user"
         }
 
-        assertEquals(0.8, updatedPrompt.params.temperature)
-        assertEquals(newSpeculation, updatedPrompt.params.speculation)
-        assertTrue(updatedPrompt.params.schema is LLMParams.Schema.JSON.Full)
-        assertEquals(schemaName, updatedPrompt.params.schema?.name)
-        assertTrue(updatedPrompt.params.toolChoice is LLMParams.ToolChoice.Required)
-        assertEquals("updated_user", updatedPrompt.params.user)
-    }
-
-    @Test
-    fun testEmptyPrompt() {
-        val emptyPrompt = Prompt.Empty
-
-        assertTrue(emptyPrompt.messages.isEmpty())
-        assertEquals("", emptyPrompt.id)
-        assertEquals(LLMParams(), emptyPrompt.params)
-
-        val json = Json.encodeToString(emptyPrompt)
-        val decoded = Json.decodeFromString<Prompt>(json)
-
-        assertEquals(emptyPrompt, decoded)
-        assertTrue(decoded.messages.isEmpty())
-        assertEquals("", decoded.id)
-    }
-
-    @Test
-    fun testPromptWithEmptyMessages() {
-        val prompt = Prompt(emptyList(), promptId)
-
-        assertTrue(prompt.messages.isEmpty())
-        assertEquals(promptId, prompt.id)
-
-        val json = Json.encodeToString(prompt)
-        val decoded = Json.decodeFromString<Prompt>(json)
-
-        assertEquals(prompt, decoded)
-        assertTrue(decoded.messages.isEmpty())
-    }
-
-    @Test
-    fun testMessageWithEmptyContent() {
-        val emptySystemMessage = Message.System("", testReqMetaInfo)
-        val emptyUserMessage = Message.User("", testReqMetaInfo)
-        val emptyAssistantMessage = Message.Assistant("", testRespMetaInfo)
-        val emptyToolCallMessage = Message.Tool.Call(toolCallId, toolName, "", testRespMetaInfo)
-        val emptyToolResultMessage = Message.Tool.Result(toolCallId, toolName, "", testReqMetaInfo)
-
-        assertEquals("", emptySystemMessage.content)
-        assertEquals("", emptyUserMessage.content)
-        assertEquals("", emptyAssistantMessage.content)
-        assertEquals("", emptyToolCallMessage.content)
-        assertEquals("", emptyToolResultMessage.content)
-
-        val prompt = Prompt.build(promptId) {
-            system("")
-            user("")
-            assistant("")
-            tool {
-                call(toolCallId, toolName, "")
-                result(toolCallId, toolName, "")
+        assertSoftly(updatedPrompt.params) {
+            temperature shouldBe 0.8
+            this.speculation shouldBe newSpeculation
+            withClue("Schema should be of type Full") {
+                schema.shouldBeInstanceOf<LLMParams.Schema.JSON.Full>()
             }
-        }
-
-        assertEquals(5, prompt.messages.size)
-        prompt.messages.forEach { message ->
-            assertEquals("", message.content)
-        }
-
-        val json = Json.encodeToString(prompt)
-        val decoded = Json.decodeFromString<Prompt>(json)
-
-        assertEquals(prompt, decoded)
-        decoded.messages.forEach { message ->
-            assertEquals("", message.content)
-        }
-    }
-
-    @Test
-    fun testToolMessagesWithNullId() {
-        val toolCallWithNullId = Message.Tool.Call(null, toolName, toolCallContent, testRespMetaInfo)
-        val toolResultWithNullId = Message.Tool.Result(null, toolName, toolCallContent, testReqMetaInfo)
-
-        assertNull(toolCallWithNullId.id)
-        assertNull(toolResultWithNullId.id)
-
-        val prompt = Prompt.build(promptId) {
-            tool {
-                call(null, toolName, toolCallContent)
-                result(null, toolName, toolCallContent)
+            schema?.name shouldBe schemaName
+            withClue("ToolChoice should be Required") {
+                toolChoice.shouldBeInstanceOf<LLMParams.ToolChoice.Required>()
             }
+            user shouldBe "updated_user"
         }
-
-        assertEquals(2, prompt.messages.size)
-
-        assertTrue(prompt.messages[0] is Message.Tool.Call)
-        assertTrue(prompt.messages[1] is Message.Tool.Result)
-        assertNull((prompt.messages[0] as Message.Tool.Call).id)
-        assertNull((prompt.messages[1] as Message.Tool.Result).id)
-
-        val json = Json.encodeToString(prompt)
-        val decoded = Json.decodeFromString<Prompt>(json)
-
-        assertEquals(prompt, decoded)
-        assertNull((decoded.messages[0] as Message.Tool.Call).id)
-        assertNull((decoded.messages[1] as Message.Tool.Result).id)
-    }
-
-    @Test
-    fun testAssistantMessageWithNullFinishReason() {
-        val prompt = Prompt.build(promptId) {
-            message(Message.Assistant(assistantMessage, testRespMetaInfo, null))
-        }
-
-        assertEquals(1, prompt.messages.size)
-        assertTrue(prompt.messages[0] is Message.Assistant)
-        assertNull((prompt.messages[0] as Message.Assistant).finishReason)
-
-        val json = Json.encodeToString(prompt)
-        val decoded = Json.decodeFromString<Prompt>(json)
-
-        assertEquals(prompt, decoded)
-        assertNull((decoded.messages[0] as Message.Assistant).finishReason)
-    }
-
-    @Test
-    fun testInvalidToolCallJsonContent() {
-        val toolCallWithInvalidJson = Message.Tool.Call(toolCallId, toolName, "invalid json", testRespMetaInfo)
-
-        assertThrows<SerializationException> {
-            toolCallWithInvalidJson.contentJson
-        }
-    }
-
-    @Test
-    fun testLLMParamsWithNullValues() {
-        val params = LLMParams(
-            temperature = null,
-            speculation = null,
-            schema = null,
-            toolChoice = null,
-            user = null
-        )
-
-        val prompt = Prompt(emptyList(), promptId, params)
-
-        assertNull(prompt.params.temperature)
-        assertNull(prompt.params.speculation)
-        assertNull(prompt.params.schema)
-        assertNull(prompt.params.toolChoice)
-        assertNull(prompt.params.user)
-
-        val json = Json.encodeToString(prompt)
-        val decoded = Json.decodeFromString<Prompt>(json)
-
-        assertEquals(prompt, decoded)
-        assertNull(decoded.params.temperature)
-        assertNull(decoded.params.speculation)
-        assertNull(decoded.params.schema)
-        assertNull(decoded.params.toolChoice)
-        assertNull(decoded.params.user)
-    }
-
-    @Test
-    fun testToolChoiceNamedWithEmptyName() {
-        val toolChoiceWithEmptyName = LLMParams.ToolChoice.Named(emptyName)
-        val prompt = basicPrompt.withUpdatedParams {
-            toolChoice = toolChoiceWithEmptyName
-        }
-
-        assertTrue(prompt.params.toolChoice is LLMParams.ToolChoice.Named)
-        assertEquals(emptyName, (prompt.params.toolChoice as LLMParams.ToolChoice.Named).name)
-
-        val json = Json.encodeToString(prompt)
-        val decoded = Json.decodeFromString<Prompt>(json)
-
-        assertEquals(prompt, decoded)
-        assertTrue(decoded.params.toolChoice is LLMParams.ToolChoice.Named)
-        assertEquals(emptyName, (decoded.params.toolChoice as LLMParams.ToolChoice.Named).name)
-    }
-
-    @Test
-    fun testSchemaWithEmptyName() {
-        val schemaWithEmptyName = LLMParams.Schema.JSON.Simple(
-            emptyName,
-            buildJsonObject { put("type", "string") }
-        )
-
-        val prompt = basicPrompt.withUpdatedParams {
-            schema = schemaWithEmptyName
-        }
-
-        assertTrue(prompt.params.schema is LLMParams.Schema.JSON.Simple)
-        assertEquals(emptyName, prompt.params.schema?.name)
-
-        val json = Json.encodeToString(prompt)
-        val decoded = Json.decodeFromString<Prompt>(json)
-
-        assertEquals(prompt, decoded)
-        assertTrue(decoded.params.schema is LLMParams.Schema.JSON.Simple)
-        assertEquals(emptyName, decoded.params.schema?.name)
-    }
-
-    @Test
-    fun testToolMessagesWithEmptyToolName() {
-        val toolCallWithEmptyName = Message.Tool.Call(toolCallId, emptyName, toolCallContent, testRespMetaInfo)
-        val toolResultWithEmptyName = Message.Tool.Result(toolCallId, emptyName, toolCallContent, testReqMetaInfo)
-
-        val prompt = Prompt.build(promptId) {
-            tool {
-                call(toolCallWithEmptyName)
-                result(toolResultWithEmptyName)
-            }
-        }
-
-        assertEquals(2, prompt.messages.size)
-        assertTrue(prompt.messages[0] is Message.Tool.Call)
-        assertTrue(prompt.messages[1] is Message.Tool.Result)
-        assertEquals(toolCallWithEmptyName.tool, (prompt.messages[0] as Message.Tool.Call).tool)
-        assertEquals(toolResultWithEmptyName.tool, (prompt.messages[1] as Message.Tool.Result).tool)
-
-        val json = Json.encodeToString(prompt)
-        val decoded = Json.decodeFromString<Prompt>(json)
-
-        assertEquals(prompt, decoded)
-        assertEquals(
-            (prompt.messages[0] as Message.Tool.Call).tool,
-            (decoded.messages[0] as Message.Tool.Call).tool
-        )
-        assertEquals(
-            (prompt.messages[1] as Message.Tool.Result).tool,
-            (decoded.messages[1] as Message.Tool.Result).tool
-        )
-    }
-
-    @Test
-    fun testLLMParamsWithExtremeTemperatureValues() {
-        val lowTemp = -1.0
-        val highTemp = 100.0
-
-        val promptWithNegativeTemp = basicPrompt.withUpdatedParams {
-            temperature = lowTemp
-        }
-        val promptWithHighTemp = basicPrompt.withUpdatedParams {
-            temperature = highTemp
-        }
-
-        assertEquals(lowTemp, promptWithNegativeTemp.params.temperature)
-        assertEquals(highTemp, promptWithHighTemp.params.temperature)
-
-        val jsonNegative = Json.encodeToString(promptWithNegativeTemp)
-        val jsonHigh = Json.encodeToString(promptWithHighTemp)
-
-        val decodedNegative = Json.decodeFromString<Prompt>(jsonNegative)
-        val decodedHigh = Json.decodeFromString<Prompt>(jsonHigh)
-
-        assertEquals(promptWithNegativeTemp.params.temperature, decodedNegative.params.temperature)
-        assertEquals(promptWithHighTemp.params.temperature, decodedHigh.params.temperature)
-    }
-
-    @Test
-    fun testSchemaWithEmptyJsonObject() {
-        val emptySchemaName = "empty-schema"
-        val emptyJsonSchema = buildJsonObject { }
-
-        val schemaWithEmptyJson = LLMParams.Schema.JSON.Simple(emptySchemaName, emptyJsonSchema)
-
-        assertTrue(schemaWithEmptyJson.schema.entries.isEmpty())
-
-        val prompt = basicPrompt.withUpdatedParams {
-            schema = schemaWithEmptyJson
-        }
-
-        assertTrue(prompt.params.schema is LLMParams.Schema.JSON.Simple)
-        assertEquals(emptySchemaName, prompt.params.schema?.name)
-        assertTrue((prompt.params.schema as LLMParams.Schema.JSON.Simple).schema.entries.isEmpty())
-
-        val json = Json.encodeToString(prompt)
-        val decoded = Json.decodeFromString<Prompt>(json)
-
-        assertEquals(prompt, decoded)
-        assertTrue(decoded.params.schema is LLMParams.Schema.JSON.Simple)
-        assertEquals(emptySchemaName, decoded.params.schema?.name)
-        assertTrue((decoded.params.schema as LLMParams.Schema.JSON.Simple).schema.entries.isEmpty())
     }
 
     @Test
@@ -594,17 +151,17 @@ class PromptTest {
             messages + Message.Assistant("How can I help you?", testRespMetaInfo)
         }
 
-        assertNotEquals(originalPrompt, updatedPrompt)
-        assertEquals(3, updatedPrompt.messages.size)
-        assertEquals(Message.Assistant("How can I help you?", testRespMetaInfo), updatedPrompt.messages[2])
+        updatedPrompt shouldNotBe originalPrompt
+        updatedPrompt.messages.size shouldBe 3
+        updatedPrompt.messages[2] shouldBe Message.Assistant("How can I help you?", testRespMetaInfo)
 
         // Test replacing messages
         val replacedPrompt = originalPrompt.withMessages {
             listOf(Message.System("You are a coding assistant", testReqMetaInfo))
         }
 
-        assertEquals(1, replacedPrompt.messages.size)
-        assertEquals(Message.System("You are a coding assistant", testReqMetaInfo), replacedPrompt.messages[0])
+        replacedPrompt.messages.size shouldBe 1
+        replacedPrompt.messages[0] shouldBe Message.System("You are a coding assistant", testReqMetaInfo)
     }
 
     @Test
@@ -621,11 +178,14 @@ class PromptTest {
 
         val updatedPrompt = originalPrompt.withParams(newParams)
 
-        assertNotEquals(originalPrompt, updatedPrompt)
-        assertEquals(newParams, updatedPrompt.params)
-        assertEquals(0.7, updatedPrompt.params.temperature)
-        assertEquals("test speculation", updatedPrompt.params.speculation)
-        assertEquals("test_user", updatedPrompt.params.user)
+        updatedPrompt shouldNotBe originalPrompt
+        updatedPrompt.params shouldBe newParams
+
+        assertSoftly(updatedPrompt.params) {
+            temperature shouldBe 0.7
+            speculation shouldBe "test speculation"
+            user shouldBe "test_user"
+        }
     }
 
     @Test
@@ -639,8 +199,8 @@ class PromptTest {
             temperature = 0.8
         }
 
-        assertNotEquals(originalPrompt, tempUpdatedPrompt)
-        assertEquals(0.8, tempUpdatedPrompt.params.temperature)
+        tempUpdatedPrompt shouldNotBe originalPrompt
+        tempUpdatedPrompt.params.temperature shouldBe 0.8
 
         // Test updating multiple parameters
         val multiUpdatedPrompt = originalPrompt.withUpdatedParams {
@@ -650,9 +210,11 @@ class PromptTest {
             user = "new_user"
         }
 
-        assertEquals(0.5, multiUpdatedPrompt.params.temperature)
-        assertEquals("new speculation", multiUpdatedPrompt.params.speculation)
-        assertEquals(LLMParams.ToolChoice.Auto, multiUpdatedPrompt.params.toolChoice)
-        assertEquals("new_user", multiUpdatedPrompt.params.user)
+        assertSoftly(multiUpdatedPrompt.params) {
+            temperature shouldBe 0.5
+            speculation shouldBe "new speculation"
+            toolChoice shouldBe LLMParams.ToolChoice.Auto
+            user shouldBe "new_user"
+        }
     }
 }
